+++ src/server/security-server-cookie.c
@@ -28,7 +28,9 @@
 #include <fcntl.h>
 #include <sys/smack.h>
 
-#include "security-server-cookie.h"
+#include <security-server-cookie.h>
+#include <security-server-comm.h>
+#include <smack-check.h>
 
 /* Delete useless cookie item *
  * then connect prev and next */
@@ -146,7 +148,8 @@ cookie_list *search_existing_cookie(int pid, const cookie_list *c_list)
 				return NULL;
 			}
 			/* Check the path is different */
-			if(strncmp(cmdline, current->path, current->path_len) != 0 || strlen(cmdline) != current->path_len)
+			if(strncmp(cmdline, current->path, current->path_len) != 0
+                || (int)strlen(cmdline) != current->path_len)
 			{
 				SEC_SVR_DBG("pid [%d] has been reused by %s. deleting the old cookie.", pid, cmdline);
 				debug_cmdline = malloc(current->path_len + 1);
